[MIG] payment_redsys: Migration to v15.0
