+++ payment_redsys/__manifest__.py
@@ -5,13 +5,12 @@
     "name": "Pasarela de pago Redsys",
     "category": "Payment Acquirer",
     "summary": "Payment Acquirer: Redsys Implementation",
-    "version": "14.0.2.0.4",
+    "version": "15.0.1.0.0",
     "author": "Tecnativa, Odoo Community Association (OCA)",
     "website": "https://github.com/OCA/l10n-spain",
     "depends": ["payment", "website_sale"],
     "external_dependencies": {"python": ["pycryptodome"]},
     "data": [
-        "views/redsys.xml",
         "views/payment_acquirer.xml",
         "views/payment_redsys_templates.xml",
         "data/payment_redsys.xml",
