+++ payment_redsys/models/__init__.py
@@ -1,1 +1,3 @@
-from . import redsys
+from . import account_payment_method
+from . import payment_acquirer
+from . import payment_transaction
