+++ payment_redsys/readme/CONTRIBUTORS.rst
@@ -10,3 +10,7 @@
 * `Acysos S.L. <https://www.acysos.com>`_:
 
   * Ignacio Ibeas <ignacio@acysos.com>
+
+* `Planes Asesoria y Soluciones Informaticas  S.L. <https://www.planesnet.com>`_:
+
+  * Luis Planes <luis@planesnet.com>
