+++ payment_redsys/tests/test_redsys.py
@@ -1,333 +1,160 @@
 # Copyright 2016-2017 Tecnativa - Sergio Teruel
+# Copyright 2023 Planesnet - Luis Planes, Laia Espinosa, Raul Solana
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 
 import json
 import logging
-from unittest.mock import patch
 
-from lxml import objectify
+from odoo.exceptions import ValidationError
+from odoo.tests import tagged
+from odoo.tools import mute_logger
 
-from odoo import http
-from odoo.tests.common import HttpCase
+from .common import RedsysCommon
 
 _logger = logging.getLogger(__name__)
 
 
-class RedsysCase(HttpCase):
-    def setUp(self):
-        super(RedsysCase, self).setUp()
-        self.currency_euro = self.env["res.currency"].search(
-            [("name", "=", "EUR")], limit=1
+@tagged("post_install", "-at_install")
+class RedsysTest(RedsysCommon):
+    def _get_expected_values(self, reference=None):
+        if not reference:
+            reference = self.reference
+        values = {
+            "reference": reference,
+            "amount": self.amount,
+        }
+        rendering_values = self._get_tx(reference)._get_specific_rendering_values(
+            values
         )
-        self.country_spain = self.env["res.country"].search(
-            [("code", "=", "ES")], limit=1
-        )
-        self.buyer_values = {
-            "partner_name": "Norbert Buyer",
-            "partner_lang": "en_US",
-            "partner_email": "norbert.buyer@example.com",
-            "partner_address": "Huge Street 2/543",
-            "partner_phone": "0032 12 34 56 78",
-            "partner_city": "Sin City",
-            "partner_zip": "1000",
-            "partner_country": self.country_spain.id,
-            "partner_country_id": self.country_spain.id,
-            "partner_country_name": "Belgium",
-            "billing_partner_name": "Norbert Buyer",
-            "billing_partner_lang": "en_US",
-            "billing_partner_email": "norbert.buyer@example.com",
-            "billing_partner_address": "Huge Street 2/543",
-            "billing_partner_phone": "0032 12 34 56 78",
-            "billing_partner_city": "Sin City",
-            "billing_partner_zip": "1000",
-            "billing_partner_country": self.country_spain.id,
-            "billing_partner_country_id": self.country_spain.id,
-            "billing_partner_country_name": "Belgium",
+        return {
+            "data_set": None,
+            "Ds_SignatureVersion": rendering_values["Ds_SignatureVersion"],
+            "Ds_MerchantParameters": rendering_values["Ds_MerchantParameters"],
+            "Ds_Signature": rendering_values["Ds_Signature"],
         }
 
-        # test partner
-        self.buyer = self.env["res.partner"].create(
-            {
-                "name": "Norbert Buyer",
-                "lang": "en_US",
-                "email": "norbert.buyer@example.com",
-                "street": "Huge Street",
-                "street2": "2/543",
-                "phone": "0032 12 34 56 78",
-                "city": "Sin City",
-                "zip": "1000",
-                "country_id": self.country_spain.id,
-            }
+    def test_compatible_acquirers(self):
+        acquirers = self.env["payment.acquirer"]._get_compatible_acquirers(
+            partner_id=self.partner.id,
+            currency_id=self.currency_euro.id,
+            company_id=self.company.id,
         )
-        self.buyer_id = self.buyer.id
-        self.redsys = self.env.ref("payment_redsys.payment_acquirer_redsys")
-        bank_journal = self.env["account.journal"].search(
-            [("type", "=", "bank")], limit=1
+        self.assertIn(self.redsys, acquirers)
+
+    def test_redirect_form_values(self):
+        tx = self.create_transaction(flow="redirect", reference="Valid transaction")
+        expected_values = self._get_expected_values(tx.reference)
+
+        with mute_logger("odoo.addons.payment.models.payment_transaction"):
+            processing_values = tx._get_processing_values()
+
+        form_info = self._extract_values_from_html_form(
+            processing_values["redirect_form_html"]
         )
-        self.redsys.journal_id = bank_journal.id
-        self.redsys.redsys_merchant_code = "069611024"
-        self.redsys.redsys_secret_key = "sq7HjrUOBfKmC576ILgskD5srU870gJ8"
-        self.redsys_ds_parameters = {
+
+        self.assertEqual(
+            form_info["action"], "https://sis-t.redsys.es:25443/sis/realizarPago/"
+        )
+        self.assertDictEqual(
+            expected_values,
+            form_info["inputs"],
+            "Redsys: invalid inputs specified in the redirect form.",
+        )
+
+    def _prepare_post_data(self, values):
+        # Simulate data received from Redsys
+        merchant_parameters = self.redsys._url_encode64(json.dumps(values))
+
+        return {
+            "Ds_MerchantParameters": merchant_parameters,
+            "Ds_Signature": self.redsys.sign_parameters(
+                self.redsys.redsys_secret_key, merchant_parameters.decode("utf8")
+            ),
+        }
+
+    def test_process_feedback_data(self):
+        tx = self.create_transaction(flow="redirect", reference="Valid transaction")
+        values = {
+            "Ds_Order": tx.reference,
             "Ds_AuthorisationCode": "999999",
-            "Ds_Date": "14%2F05%2F2017",
-            "Ds_Card_Brand": "1",
-            "Ds_SecurePayment": "1",
-            "Ds_MerchantData": "xxx",
-            "Ds_Card_Country": "724",
-            "Ds_Terminal": "001",
-            "Ds_MerchantCode": "069611024",
-            "Ds_ConsumerLanguage": "1",
-            "Ds_Response": "0000",
-            "Ds_Order": "TST0001",
-            "Ds_Currency": "978",
-            "Ds_Amount": "10050",
-            "Ds_TransactionType": "0",
-            "Ds_Hour": "08%3A29",
+            "Ds_Response": "100",
         }
-        self.vals_tx = {
-            "amount": 100.50,
-            "acquirer_id": self.redsys.id,
-            "currency_id": self.currency_euro.id,
-            "reference": "TST0001",
-            "partner_id": self.buyer_id,
+        post_data = self._prepare_post_data(values)
+
+        tx = self.env["payment.transaction"]._get_tx_from_feedback_data(
+            "redsys", post_data
+        )
+        tx._process_feedback_data(post_data)
+        self.assertEqual(
+            tx.state, "done", "Redsys: validation did not put tx into done state"
+        )
+
+    def test_unknown_transaction(self):
+        # typical data posted by Redsys after client has successfully paid
+        # unknown transaction
+        values = {
+            "Ds_Order": "unknown transaction",
+            "Ds_AuthorisationCode": "999999",
         }
-        self.tx = self.env["payment.transaction"].create(self.vals_tx)
-        self.partner = self.env["res.partner"].create({"name": "Partner Test"})
-        self.product = self.env["product.product"].create(
-            {"name": "Test Product", "list_price": 100.50}
-        )
-        self.so = self.env["sale.order"].create(
-            {
-                "partner_id": self.partner.id,
-                "partner_invoice_id": self.partner.id,
-                "partner_shipping_id": self.partner.id,
-                "order_line": [
-                    (
-                        0,
-                        0,
-                        {
-                            "name": "Test",
-                            "product_id": self.product.id,
-                            "product_uom_qty": 1,
-                            "product_uom": self.product.uom_id.id,
-                            "price_unit": self.product.list_price,
-                            "tax_id": [(6, 0, [])],
-                        },
-                    )
-                ],
-                "pricelist_id": self.env.ref("product.list0").id,
-            }
-        )
-        merchant_parameters = self.redsys._prepare_merchant_parameters(self.vals_tx)
-        self.ds_signature = self.redsys.sign_parameters(
-            self.redsys.redsys_secret_key, merchant_parameters
-        )
+        post_data = self._prepare_post_data(values)
+        with self.assertRaises(ValidationError):
+            self.env["payment.transaction"]._handle_feedback_data("redsys", post_data)
 
-    def data_post_redsys(self, url=None, data=None, timeout=60):
-        doc = self.url_open(url, data=data, timeout=timeout)
-        return doc
+    def test_feedback_processing(self):
+        # typical data posted by Redsys after client has successfully paid
 
-    def test_10_redsys_form_render(self):
-        # be sure not to do stupid things
-        self.assertNotEqual(
-            self.redsys.state, "enabled", "test without test environment"
-        )
-        base_url = self.env["ir.config_parameter"].get_param("web.base.url")
-
-        # ----------------------------------------
-        # Test: button direct rendering
-        # ----------------------------------------
-
-        # render the button
-        res = self.redsys.render(
-            "TST0001",
-            100.50,
-            self.currency_euro.id,
-            partner_id=None,
-            values=self.buyer_values,
-        )
-
-        # check form result
-        tree = objectify.fromstring(res)
-        self.assertEqual(
-            tree.input.attrib.get("data-action-url"),
-            "https://sis-t.redsys.es:25443/sis/realizarPago/",
-            "Redsys: wrong form POST url",
-        )
-
-        for form_input in tree.input:
-            if form_input.get("name") in ["submit"]:
-                continue
-            if form_input.get("name") == "Ds_MerchantParameters":
-                DS_parameters = self.redsys._url_decode64(form_input.get("value"))
-                self.assertEqual(DS_parameters["Ds_Merchant_Order"], "TST0001")
-                self.assertEqual(
-                    DS_parameters["Ds_Merchant_MerchantUrl"],
-                    "%s/payment/redsys/return" % base_url,
-                )
-        # Form values
-        values = self.redsys.redsys_form_generate_values(self.vals_tx)
-        self.assertEqual(
-            values["Ds_Signature"], self.ds_signature, "Redsys: Sign failed"
-        )
-
-    def test_error(self):
-        self.data_post_redsys(url="/payment/redsys/return")
-
-    def _form_feedback(self, post_data):
-        # This method is created to simulate what the controller does avoiding
-        # the petition.
-        self.env["payment.transaction"].form_feedback(post_data, "redsys")
-
-    def test_20_redsys_form_management(self):
-        # be sure not to do stupid thing
-        self.assertNotEqual(
-            self.redsys.state, "enabled", "test without test environment"
-        )
-        DS_parameters = self.redsys._url_encode64(json.dumps(self.redsys_ds_parameters))
-        # typical data posted by redsys after client has successfully paid
-        redsys_post_data = {
-            "Ds_Signature": self.redsys.sign_parameters(
-                self.redsys.redsys_secret_key, DS_parameters
-            ),
-            "Ds_MerchantParameters": DS_parameters,
-            "Ds_SignatureVersion": "HMAC_SHA256_V1",
+        # redsys not authorisation code
+        values = {
+            "Ds_Order": "Valid transaction",
+            "Ds_AuthorisationCode": "",
         }
-        self._form_feedback(redsys_post_data)
-        self.assertEqual(self.tx.redsys_txnid, "999999", "Redsys: Get transaction")
-        self.assertEqual(self.tx.reference, "TST0001", "Redsys: Get transaction")
-        # check state
-        self.assertEqual(
-            self.tx.state, "done", "Redsys: validation did not put tx into done state"
-        )
-        self.assertEqual(
-            self.tx.redsys_txnid,
-            self.redsys_ds_parameters.get("Ds_AuthorisationCode"),
-            "Redsys: validation did not update tx payid",
-        )
+        post_data = self._prepare_post_data(values)
+        with self.assertRaises(ValidationError):
+            self.env["payment.transaction"]._handle_feedback_data("redsys", post_data)
 
-    def test_30_redsys_form_management(self):
-        # be sure not to do stupid thing
-        self.assertNotEqual(
-            self.redsys.state, "enabled", "test without test environment"
-        )
-        unknow_tx = self.redsys_ds_parameters.copy()
-        unknow_tx["Ds_Order"] = "22"
-        DS_parameters = self.redsys._url_encode64(json.dumps(unknow_tx))
-        # typical data posted by redsys after client has successfully paid
-        redsys_post_data = {
-            "Ds_Signature": self.redsys.sign_parameters(
-                self.redsys.redsys_secret_key, DS_parameters
-            ),
-            "Ds_MerchantParameters": DS_parameters,
-            "Ds_SignatureVersion": "HMAC_SHA256_V1",
+        # Valid transaction. status: done
+        tx = self.create_transaction(flow="redirect", reference="Valid transaction")
+        values = {
+            "Ds_Order": tx.reference,
+            "Ds_AuthorisationCode": "999999",
+            "Ds_Response": "100",
         }
-        self._form_feedback(redsys_post_data)
+        post_data = self._prepare_post_data(values)
+
+        tx._handle_feedback_data("redsys", post_data)
         self.assertEqual(
-            http.OpenERPSession.tx_error, True, "test with unknow transacction"
+            tx.state, "done", "Redsys: validation did not put tx into done state"
         )
 
-    def test_40_redsys_form_error(self):
-        # be sure not to do stupid thing
-        self.assertNotEqual(
-            self.redsys.state, "enabled", "test without test environment"
-        )
-        # Force error from provider
-        error_tx = self.redsys_ds_parameters.copy()
-        error_tx["Ds_Response"] = "9065"
-        DS_parameters = self.redsys._url_encode64(json.dumps(error_tx))
-        # typical data posted by redsys after client has successfully paid
-        redsys_post_data = {
-            "Ds_Signature": self.redsys.sign_parameters(
-                self.redsys.redsys_secret_key, DS_parameters
-            ),
-            "Ds_MerchantParameters": DS_parameters,
-            "Ds_SignatureVersion": "HMAC_SHA256_V1",
+        # No valid card transaction. status: pending
+        tx = self.create_transaction(flow="redirect", reference="Pending transaction")
+        values = {
+            "Ds_Order": tx.reference,
+            "Ds_AuthorisationCode": "999999",
+            "Ds_Response": "203",
         }
-        self._form_feedback(redsys_post_data)
-        # Get transaction
-        self.assertNotEqual(
-            self.tx.state, "done", "Redsys: validation did not put tx into done state"
-        )
+        post_data = self._prepare_post_data(values)
+        tx._handle_feedback_data("redsys", post_data)
+        self.assertEqual(tx.state, "pending", "Redsys: pending transaction status")
 
-    def test_50_redsys_form_pending(self):
-        # be sure not to do stupid thing
-        self.assertNotEqual(
-            self.redsys.state, "enabled", "test without test environment"
-        )
-        # Force error from provider
-        error_tx = self.redsys_ds_parameters.copy()
-        error_tx["Ds_Response"] = "912"
-        DS_parameters = self.redsys._url_encode64(json.dumps(error_tx))
-        # typical data posted by redsys after client has successfully paid
-        redsys_post_data = {
-            "Ds_Signature": self.redsys.sign_parameters(
-                self.redsys.redsys_secret_key, DS_parameters
-            ),
-            "Ds_MerchantParameters": DS_parameters,
-            "Ds_SignatureVersion": "HMAC_SHA256_V1",
+        # Cancel status
+        tx = self.create_transaction(flow="redirect", reference="Cancel transaction")
+        values = {
+            "Ds_Order": tx.reference,
+            "Ds_AuthorisationCode": "999999",
+            "Ds_Response": "913",
         }
-        self._form_feedback(redsys_post_data)
-        # Get transaction
-        self.assertEqual(
-            self.tx.state, "cancel", "Redsys: validation did not put tx into done state"
-        )
+        post_data = self._prepare_post_data(values)
+        tx._handle_feedback_data("redsys", post_data)
+        self.assertEqual(tx.state, "cancel", "Redsys: 913-9912 generic invalid card")
 
-    def test_80_redsys_partial_payment(self):
-        # be sure not to do stupid thing
-        self.assertNotEqual(
-            self.redsys.state, "enabled", "test without test environment"
-        )
-        self.redsys.redsys_percent_partial = 50
-        # Form values
-        values = self.redsys.redsys_form_generate_values(self.vals_tx)
-        Ds_MerchantParameters = self.redsys._url_decode64(
-            values["Ds_MerchantParameters"]
-        )
-        self.assertEqual(
-            Ds_MerchantParameters["Ds_Merchant_Amount"],
-            "5025",
-            "Redsys: Partial amount Ok",
-        )
-
-    @patch("odoo.addons.sale.models.sale.SaleOrder.action_confirm")
-    @patch("odoo.addons.sale.models.sale.SaleOrder.action_quotation_send")
-    def test_90_redsys_form_partial_payment(self, mock_quo_send, mock_confirm):
-        # be sure not to do stupid thing
-        self.assertNotEqual(
-            self.redsys.state, "enabled", "test without test environment"
-        )
-        self.redsys.redsys_percent_partial = 50
-        params = self.redsys_ds_parameters.copy()
-        params["Ds_Amount"] = "5025"
-        DS_parameters = self.redsys._url_encode64(json.dumps(params))
-        # typical data posted by redsys after client has successfully paid
-        redsys_post_data = {
-            "Ds_Signature": self.redsys.sign_parameters(
-                self.redsys.redsys_secret_key, DS_parameters
-            ),
-            "Ds_MerchantParameters": DS_parameters,
-            "Ds_SignatureVersion": "HMAC_SHA256_V1",
+        # Error transction status
+        tx = self.create_transaction(flow="redirect", reference="Error transaction")
+        values = {
+            "Ds_Order": tx.reference,
+            "Ds_AuthorisationCode": "999999",
+            "Ds_Response": "9999",
         }
-        self.tx.sale_order_ids = [(6, 0, self.so.ids)]
-        # Get transaction
-        self._form_feedback(redsys_post_data)
-        mock_confirm.assert_called_once_with()
-        mock_quo_send.assert_not_called()
-        # check state
-        self.assertEqual(
-            self.tx.state, "done", "Redsys: validation put tx into done state"
-        )
-        self.assertEqual(
-            self.tx.redsys_txnid,
-            self.redsys_ds_parameters.get("Ds_AuthorisationCode"),
-            "Redsys: validation did not update tx payid",
-        )
-
-    def test_91_redsys_return_post(self):
-        # be sure not to do stupid thing
-        self.assertNotEqual(
-            self.redsys.state, "enabled", "test without test environment"
-        )
-        res = self.data_post_redsys(url="/payment/redsys/return")
-        self.assertGreater(res.url.find("/shop"), 0, "Redsys: Redirection to /shop")
+        post_data = self._prepare_post_data(values)
+        tx._handle_feedback_data("redsys", post_data)
+        self.assertEqual(tx.state, "error", "Redsys: response error")
