+++ payment_redsys/views/payment_acquirer.xml
@@ -5,9 +5,9 @@
     <record id="acquirer_form_redsys" model="ir.ui.view">
         <field name="name">acquirer.form.redsys</field>
         <field name="model">payment.acquirer</field>
-        <field name="inherit_id" ref="payment.acquirer_form" />
+        <field name="inherit_id" ref="payment.payment_acquirer_form" />
         <field name="arch" type="xml">
-            <xpath expr='//group[@name="acquirer"]' position='after'>
+            <xpath expr='//group[@name="acquirer"]' position='inside'>
                 <group attrs="{'invisible': [('provider', '!=', 'redsys')]}">
                     <field
                         name="redsys_merchant_name"
@@ -53,10 +53,6 @@
                         name="redsys_signature_version"
                         attrs="{'required': [('provider', '=', 'redsys')]}"
                     />
-                    <field
-                        name="redsys_percent_partial"
-                        attrs="{'required': [('provider', '=', 'redsys')]}"
-                    />
                 </group>
             </xpath>
         </field>
@@ -64,9 +60,9 @@
     <record id="transaction_form_redsys" model="ir.ui.view">
         <field name="name">acquirer.transaction.form.redsys</field>
         <field name="model">payment.transaction</field>
-        <field name="inherit_id" ref="payment.transaction_form" />
+        <field name="inherit_id" ref="payment.payment_transaction_form" />
         <field name="arch" type="xml">
-            <field name="date" position='after'>
+            <field name="create_date" position='after'>
                 <field name="redsys_txnid" />
             </field>
         </field>
