+++ payment_redsys/views/payment_redsys_templates.xml
@@ -1,9 +1,10 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <!-- Copyright 2017 Tecnativa - Sergio Teruel
+     Copyright 2023 Planesnet - Luis Planes, Laia Espinosa, Raul Solana
      License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl-3). -->
 <odoo noupdate="0">
     <template id="redsys_form">
-        <div>
+        <form t-att-action="api_url" method="post">
             <input
                 type="hidden"
                 name="data_set"
@@ -21,6 +22,6 @@
                 t-att-value="Ds_MerchantParameters"
             />
             <input type="hidden" name="Ds_Signature" t-att-value="Ds_Signature" />
-        </div>
+        </form>
     </template>
 </odoo>
