--- payment_redsys/views/redsys.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<!-- Copyright 2017 Tecnativa - Sergio Teruel
-     License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl-3). -->
-<odoo>
-    <template id="redsys_acquirer_button">
-        <form t-if="acquirer" t-att-action="tx_url" method="post" target="_self">
-            <input
-                type="hidden"
-                name="Ds_SignatureVersion"
-                t-att-value="Ds_SignatureVersion"
-            />
-            <input
-                type="hidden"
-                name="Ds_MerchantParameters"
-                t-att-value="Ds_MerchantParameters"
-            />
-            <input type="hidden" name="Ds_Signature" t-att-value="Ds_Signature" />
-            <button type="submit" width="100px" t-att-class="submit_class">
-                <img
-                    t-if="not submit_txt"
-                    src="/payment_redsys/static/src/img/redsys_icon.png"
-                />
-                <span t-if="submit_txt">
-                    <t t-esc="submit_txt" />
-                    <span class="fa fa-long-arrow-right" />
-                </span>
-            </button>
-        </form>
-    </template>
-</odoo>
